Updating WPT template to handle tests with timeouts
The WPT test framework does not report "subtest" results if any of the
tests in a test file hang or otherwise cause the test to time out.
Accordingly, the results page for WPT results needs to take into account
that some browsers might have missing test results because of this, even
though the tests were attempted to run. The simple technique of looping
through the results map and creating the table cells won't work here; we
must introspect into the results map and create cells only for the results
we have, and create dummy cells for the ones we do not.