Adding support for "known failures"
If a test fails in a test suite, but the failure is due to an inherent
limitation of the browser or the driver, the test should still count as
failing, but it should be possible to display on the reports that such a
failure is expected. This is especially important for WPT tests, where
there is no such concept as a "skipped" test, and where tests may fail
because a browser doesn't implement a necessary API to run the test.